+++ verification/v1.9.3/love_test/conf.lua
@@ -1,0 +1,8 @@
+function love.conf(t)
+ t.identity='autobike-193-release-integration-test'
+ t.version='11.5'
+ t.window=false
+ t.modules.graphics=false
+ t.modules.joystick=false
+ t.modules.physics=false
+end